+++ Gruntfile.js
@@ -35,14 +35,14 @@ module.exports = function (grunt) {
           sourceMap: true,
           sourceMapName: 'dist/angular-cache.min.map',
           banner: '/**\n' +
-            '* @author Jason Dobry <jason.dobry@gmail.com>\n' +
-            '* @file angular-cache.min.js\n' +
-            '* @version <%= pkg.version %> - Homepage <https://github.com/jmdobry/angular-cache>\n' +
-            '* @copyright (c) 2013-2014 Jason Dobry <http://www.pseudobry.com>\n' +
-            '* @license MIT <https://github.com/jmdobry/angular-cache/blob/master/LICENSE>\n' +
-            '*\n' +
-            '* @overview angular-cache is a very useful replacement for Angular\'s $cacheFactory.\n' +
-            '*/\n'
+          '* @author Jason Dobry <jason.dobry@gmail.com>\n' +
+          '* @file angular-cache.min.js\n' +
+          '* @version <%= pkg.version %> - Homepage <https://github.com/jmdobry/angular-cache>\n' +
+          '* @copyright (c) 2013-2014 Jason Dobry <http://www.pseudobry.com>\n' +
+          '* @license MIT <https://github.com/jmdobry/angular-cache/blob/master/LICENSE>\n' +
+          '*\n' +
+          '* @overview angular-cache is a very useful replacement for Angular\'s $cacheFactory.\n' +
+          '*/\n'
         },
         files: {
           'dist/angular-cache.min.js': ['dist/angular-cache.js']
@@ -60,194 +60,18 @@ module.exports = function (grunt) {
       options: {
         configFile: './karma.conf.js'
       },
+      dist: {},
       dev: {
-        browsers: ['Chrome'],
+        browsers: ['Firefox'],
         autoWatch: true,
         singleRun: false
       },
       min: {
         options: {
           files: [
-            'bower_components/angular-1.2.16/angular.js',
-            'bower_components/angular-mocks-1.2.16/angular-mocks.js',
+            'bower_components/angular-1.2.25/angular.js',
+            'bower_components/angular-mocks-1.2.25/angular-mocks.js',
             'dist/angular-cache.min.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      min_old: {
-        options: {
-          files: [
-            'bower_components/angular-1.1.5/angular.js',
-            'bower_components/angular-mocks-1.1.5/angular-mocks.js',
-            'dist/angular-cache.min.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.0.4': {
-        options: {
-          files: [
-            'bower_components/angular-1.0.4/angular.js',
-            'bower_components/angular-mocks-1.0.4/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.0.5': {
-        options: {
-          files: [
-            'bower_components/angular-1.0.5/angular.js',
-            'bower_components/angular-mocks-1.0.5/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.0.6': {
-        options: {
-          files: [
-            'bower_components/angular-1.0.6/angular.js',
-            'bower_components/angular-mocks-1.0.6/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.0.7': {
-        options: {
-          files: [
-            'bower_components/angular-1.0.7/angular.js',
-            'bower_components/angular-mocks-1.0.7/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.0.8': {
-        options: {
-          files: [
-            'bower_components/angular-1.0.8/angular.js',
-            'bower_components/angular-mocks-1.0.8/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.1.4': {
-        options: {
-          files: [
-            'bower_components/angular-1.1.4/angular.js',
-            // hopefully this works. 1.1.4 isn't available on bower
-            'bower_components/angular-mocks-1.1.5/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.1.5': {
-        options: {
-          files: [
-            'bower_components/angular-1.1.5/angular.js',
-            'bower_components/angular-mocks-1.1.5/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.0': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.0/angular.js',
-            'bower_components/angular-mocks-1.2.0/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.1': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.1/angular.js',
-            'bower_components/angular-mocks-1.2.1/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.2': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.2/angular.js',
-            'bower_components/angular-mocks-1.2.2/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.3': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.3/angular.js',
-            'bower_components/angular-mocks-1.2.3/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.4': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.4/angular.js',
-            'bower_components/angular-mocks-1.2.4/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.5': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.5/angular.js',
-            'bower_components/angular-mocks-1.2.5/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.6': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.6/angular.js',
-            'bower_components/angular-mocks-1.2.6/angular-mocks.js',
-            'dist/angular-cache.js',
-            './karma.start.js',
-            'test/unit/**/*.js'
-          ]
-        }
-      },
-      '1.2.7': {
-        options: {
-          files: [
-            'bower_components/angular-1.2.7/angular.js',
-            'bower_components/angular-mocks-1.2.7/angular-mocks.js',
-            'dist/angular-cache.js',
             './karma.start.js',
             'test/unit/**/*.js'
           ]
@@ -269,35 +93,34 @@ module.exports = function (grunt) {
     grunt.file.write(filePath, file);
   });
 
-  grunt.registerTask('test', ['karma:1.2.7']);
+  grunt.registerTask('banner', function () {
+    var file = grunt.file.read('dist/angular-cache.js');
+
+    var banner = '/**\n' +
+      '* @author Jason Dobry <jason.dobry@gmail.com>\n' +
+      '* @file angular-cache.js\n' +
+      '* @version ' + pkg.version + ' - Homepage <https://github.com/jmdobry/angular-cache>\n' +
+      '* @copyright (c) 2013-2014 Jason Dobry <http://www.pseudobry.com>\n' +
+      '* @license MIT <https://github.com/jmdobry/angular-cache/blob/master/LICENSE>\n' +
+      '*\n' +
+      '* @overview angular-cache is a very useful replacement for Angular\'s $cacheFactory.\n' +
+      '*/\n';
+
+    file = banner + file;
+
+    grunt.file.write('dist/angular-cache.js', file);
+  });
+
+  grunt.registerTask('test', ['karma:dist']);
   grunt.registerTask('build', [
     'clean',
     'jshint',
     'browserify',
     'version:dist/angular-cache.js',
+    'banner',
     'uglify'
   ]);
   grunt.registerTask('default', ['build']);
 
-  // Used by the CLI build servers
-  grunt.registerTask('test-ci', [
-//		'karma:1.0.4',
-//		'karma:1.0.5',
-//		'karma:1.0.6',
-//		'karma:1.0.7',
-//		'karma:1.0.8',
-//		'karma:1.1.4',
-//		'karma:1.1.5',
-    'karma:1.2.1',
-    'karma:1.2.2',
-    'karma:1.2.3',
-    'karma:1.2.4',
-    'karma:1.2.5',
-    'karma:1.2.6',
-    'karma:1.2.7',
-    'karma:min',
-    'karma:min_old'
-  ]);
-  grunt.registerTask('ci', ['build', 'test-ci']);
   grunt.registerTask('go', ['build', 'watch']);
 };